feat: update lab issue 179
#179
