+++ labs/issue-179/src/index.ts
@@ -1,11 +1,11 @@
-import { orm } from 'lambdaorm'
+import { avg, orm  } from 'lambdaorm'
 import { Orders } from './northwind/domain'
 ( async ()=> {
 	try {		
 		await orm.init('lambdaORM.yaml')
 		const query = (ordNo: number) => Orders.details
-											.filter(o => o.orderId === ordNo)
-											.map(o => ({ ordNo:o.orderId, total: avg(o.unitPrice * o.quantity )}))
+											.filter( o => o.orderId === ordNo)
+											.map( o => ({ ordNo:o.orderId, total: avg(o.unitPrice * o.quantity )}))
 		const result = await orm.execute(query, { ordNo: 5 }, {stage:'MySQL'});
         console.log(JSON.stringify(result))		
 	} catch (error:any) {
